added draft editing
