+++ server/routes/posts.js
@@ -4,8 +4,8 @@ const { sendErrorResponse, sendDataResponse, sendMsgResponse, _idStrippedDoc } =
 const { authenticateToken } = require('../utils/auth')
 
 router.post('/create', authenticateToken, (req, res) => {
-  const { titleImg, title, subtitle, content, isPublished, tags } = req.body
-  addPost({ titleImg, title, subtitle, content, isPublished, tags }, (post, err) => {
+  const { title, content, isPublished, tags } = req.body
+  addPost({ title, content, isPublished, tags }, (post, err) => {
     if (err) {
       sendErrorResponse(res, 500, err)
     } else {
@@ -52,8 +52,8 @@ router.delete('', authenticateToken, (req, res) => {
 })
 
 router.put('', authenticateToken, (req, res) => {
-  const { id, titleImg, title, content, subtitle, isPublished, tags } = req.body
-  updatePost({ id, titleImg, title, content, subtitle, isPublished, tags }, (post, err) => {
+  const { id, title, content, isPublished, tags } = req.body
+  updatePost({ id, title, content, isPublished, tags }, (post, err) => {
     if (err) {
       sendErrorResponse(res, 500, err)
     } else {
