+++ src/components/posts/index.ts
@@ -1,1 +1,2 @@
 export { default as PostContent } from './PostContent'
+export { default as PostEditor } from './PostEditor'
